requirements.txt
streamlit>=1.20.0
pandas>=1.5.0
scikit-learn>=1.2.0
requests>=2.28.0
python-dotenv>=1.0.0